+++ R/backend-mysql.R
@@ -113,12 +113,17 @@ sql_translation.MariaDBConnection <- function(con) {
       # but available in MariaDB. A few more details at:
       # https://www.oreilly.com/library/view/mysql-cookbook/0596001452/ch04s11.html
       str_detect = sql_infix("REGEXP"),
-      str_like = function(string, pattern, ignore_case = TRUE) {
-        if (isTRUE(ignore_case)) {
+      str_like = function(string, pattern, ignore_case = deprecated()) {
+        ignore_case <- deprecate_ignore_case(ignore_case)
+        if (ignore_case) {
           sql_expr(!!string %LIKE% !!pattern)
         } else {
           sql_expr(!!string %LIKE BINARY% !!pattern)
         }
+      },
+      str_ilike = function(string, pattern) {
+        # MySQL's LIKE is case-insensitive by default
+        sql_expr(!!string %LIKE% !!pattern)
       },
       str_locate = function(string, pattern) {
         sql_expr(REGEXP_INSTR(!!string, !!pattern))