+++ R/backend-postgres.R
@@ -137,13 +137,16 @@ sql_translation.PqConnection <- function(con) {
         )
       },
       # https://www.postgresql.org/docs/current/functions-matching.html
-      str_like = function(string, pattern, ignore_case = TRUE) {
-        check_bool(ignore_case)
-        if (isTRUE(ignore_case)) {
+      str_like = function(string, pattern, ignore_case = deprecated()) {
+        ignore_case <- deprecate_ignore_case(ignore_case)
+        if (ignore_case) {
           sql_expr(!!string %ILIKE% !!pattern)
         } else {
           sql_expr(!!string %LIKE% !!pattern)
         }
+      },
+      str_ilike = function(string, pattern) {
+        sql_expr(!!string %ILIKE% !!pattern)
       },
       str_replace = function(string, pattern, replacement) {
         sql_expr(regexp_replace(!!string, !!pattern, !!replacement))